+++ src/Functors/Monads/ListMonad/index.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * list monad functions
+ *
+ * @package bingo-functional
+ * @author Lochemem Bruno Michael
+ * @license Apache-2.0
+ */
+
+require_once __DIR__ . '/Append.php';
+require_once __DIR__ . '/Concat.php';
+require_once __DIR__ . '/FromValue.php';
+require_once __DIR__ . '/Head.php';
+require_once __DIR__ . '/Prepend.php';
+require_once __DIR__ . '/Tail.php';
